pyproject.toml
[tool.poetry]
name = "articles"
version = "0.1.0"
description = ""
authors = ["johnpapwinter <johnpapahimonas@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.7.0"
asyncpg = "^0.30.0"
psycopg = {extras= ["binary"], version= "^3.2.3" }
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
#sqlalchemy-continuum = "^1.4.2"
elasticsearch = "^8.17.0"
elasticsearch-dsl = "^8.17.0"
aiohttp = "^3.11.11"
alembic = "^1.14.0"
bcrypt = "^4.0.1"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.25.0"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
